+++ expo/app/(tabs)/settings.tsx
@@ -41,6 +41,7 @@ export default function SettingsScreen() {
       await signOut(auth);
       await AsyncStorage.removeItem('isAuthenticated');
       Alert.alert('Logged Out', 'You have been logged out.', [{ text: 'OK' }]);
+      router.push('auth');
     } catch (error) {
       console.error('Logout Error:', error);
     }
