+++ expo/app/_layout.tsx
@@ -1,23 +1,11 @@
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import { Stack } from 'expo-router';
-import { useState, useEffect } from 'react';
+import useAuth from 'hooks/useAuth';
 import { View, ActivityIndicator } from 'react-native';
 
 import '../global.css';
 
 export default function Layout() {
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const checkAuth = async () => {
-      const authStatus = await AsyncStorage.getItem('isAuthenticated');
-      setIsAuthenticated(authStatus === 'true');
-      setLoading(false);
-    };
-
-    checkAuth();
-  }, []);
+  const { user, loading } = useAuth();
 
   if (loading) {
     return (
@@ -29,11 +17,7 @@ export default function Layout() {
 
   return (
     <Stack screenOptions={{ headerShown: false }}>
-      {!isAuthenticated ? (
-        <Stack.Screen name="auth" options={{ presentation: 'modal' }} />
-      ) : (
-        <Stack.Screen name="(tabs)" />
-      )}
+      {!user ? <Stack.Screen name="auth" /> : <Stack.Screen name="(tabs)" />}
     </Stack>
   );
 }
